refactor: move pip and npm metadata helpers to shared
